changed accession event, re-added organisms, and model refactor
