+++ core/src/schema_gnl.rs
@@ -204,6 +204,18 @@ diesel::table! {
     }
 }
 
+diesel::table! {
+    organism_entities (entity_id) {
+        entity_id -> Varchar,
+    }
+}
+
+diesel::table! {
+    accession_event_entities (entity_id) {
+        entity_id -> Varchar,
+    }
+}
+
 
 use super::schema::{
     accession_events,
